fixed printf implementation fixed int to string conversion

+++ src/kernel/lib/libc_argos.c
@@ -100,7 +100,7 @@ currently supports decimal, hex*/
     
 
     int len = 1;
-    int n0 = n;
+    long long n0 = n;
 
     const char* digits;
